Let base class handle randomness
