+++ sparse/unit_test/Test_Sparse_ccs2crs.hpp
@@ -103,9 +103,6 @@ void doAllCcs2crs(size_t m, size_t n) {
 }
 
 TEST_F(TestCategory, sparse_ccs2crs) {
-  TestUtils::initRandSeed();
-  SCOPED_TRACE("rand seed: " + std::to_string(TestUtils::getTestSeed()));
-
   // Empty cases
   doCcs2Crs<float, Kokkos::LayoutLeft, TestDevice>(1, 0, 1, 10);
   doCcs2Crs<float, Kokkos::LayoutLeft, TestDevice>(0, 1, 1, 10);
