+++ sparse/unit_test/Test_Sparse_csc2csr.hpp
@@ -94,9 +94,6 @@ void doAllCsc2csr(size_t m, size_t n) {
 }
 
 TEST_F(TestCategory, sparse_csc2csr) {
-  TestUtils::initRandSeed();
-  SCOPED_TRACE("rand seed: " + std::to_string(TestUtils::getTestSeed()));
-
   // Empty cases
   doCsc2Csr<float, Kokkos::LayoutLeft, TestDevice>(1, 0, 1, 10);
   doCsc2Csr<float, Kokkos::LayoutLeft, TestDevice>(0, 1, 1, 10);
