+++ sparse/unit_test/Test_Sparse_trsv.hpp
@@ -85,9 +85,6 @@ struct TrsvTest {
     View2D b_y("B", numCols, numMV);
     View2D b_x_copy("B", numCols, numMV);
 
-    SCOPED_TRACE("rand seed: " + std::to_string(TestUtils::getTestSeed()));
-    TestUtils::initRandSeed();
-
     Kokkos::Random_XorShift64_Pool<execution_space> rand_pool(rand());
     Kokkos::fill_random(b_x_copy, rand_pool, scalar_t(10));
