+++ test_common/Test_HIP.hpp
@@ -7,16 +7,13 @@
 #include <Kokkos_Core.hpp>
 #include <KokkosKernels_config.h>
 
+#include "Test_common.hpp"
+
 #if defined(KOKKOSKERNELS_TEST_ETI_ONLY) && !defined(KOKKOSKERNELS_ETI_ONLY)
 #define KOKKOSKERNELS_ETI_ONLY
 #endif
 
-class hip : public ::testing::Test {
- protected:
-  static void SetUpTestCase() {}
-
-  static void TearDownTestCase() {}
-};
+class hip : public CommonParent {};
 
 using HIPSpaceDevice        = Kokkos::Device<Kokkos::HIP, Kokkos::HIPSpace>;
 using HIPManagedSpaceDevice = Kokkos::Device<Kokkos::HIP, Kokkos::HIPManagedSpace>;
